chore: board 패키지로 이동
